+++ changelog.md
@@ -4,6 +4,7 @@
 
 * Fix deleted options that appear in the command line JSON file export.
 * Fix subtracting date intervals from dates in formulas in some situations not working.
+* Added length is lower than filter.
 
 ## Released (2021-11-25)